requirements.txt
psycopg2
SQLAlchemy
djangorestframework
django-extensions
pygments
django-filter